Extract useFocusedBlock hook
